final alignments for flask-rpi comms
