--- tests/tests_libs/__init__.py
@@ -1,1 +1,0 @@
-from .rpi_test_client import RpiTestClient
